Money update

+++ Assets/PC/Scripts/ShopProduct_item.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New Shop Product", menuName = "Items/Shop Product")]
 public class ShopProduct_Item : ItemBase
 {
-    [field: SerializeField] public float Price { get; private set; }
+    [field: SerializeField] public int Price { get; private set; }
     [field: SerializeField] public Sprite ProductImage { get; private set; }
     [field: SerializeField] public GameObject Prefab { get; private set; }
     [field: SerializeField] public bool Stackable { get; private set; }
